+++ src/variable_mesh.ts
