+++ src/widgyts.ts
@@ -4,5 +4,5 @@ export const _yt_tools = import('@data-exp-lab/yt-tools');
 export * from './utils';
 export * from './colormap_container';
 export * from './fixed_res_buffer';
-export * from './VariableMeshModel';
+export * from './variable_mesh';
 export * from './widgyts_canvas';
